Extract model loading into another file
